requirements.txt
streamlit==1.30.0
SQLAlchemy==2.0.20
pandas==2.1.2
plotly==5.16.0
python-dotenv==1.0.0
werkzeug==2.3.7